feat: add tests to refund event and type fixes
